Refactor: Consolidate ShopApp implementations
Merge ShopApp and ShopAppp into a single ShopApp.js file.
