+++ shop/frontend/src/ShopApp.js
@@ -17,7 +17,7 @@ import { LoginModal } from './components/LoginModal';
 import { StoreBanner } from './components/StoreBanner';
 
 const Main = ({ siteSlug = 'default', initialCategoryId }) => {
-  const { state, setFulfillmentType, addItem } = useCart();
+  const { state, setFulfillmentType, addItem, getCartTotal } = useCart();
   const [privacyOpen, setPrivacyOpen] = useState(true);
   const [fulfillmentOpen, setFulfillmentOpen] = useState(false);
 
@@ -33,9 +33,8 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
   const [vegFilter, setVegFilter] = useState('all');
   const [lastDeliveryId, setLastDeliveryId] = useState(null);
 
-  // State for order type selection
-  const [orderType, setOrderType] = useState<'pickup' | 'delivery'>('pickup');
-  const [pickupTime, setPickupTime] = useState('10:00 AM');
+  // Additional UI state brought from the alternate implementation
+  const [pickupTime] = useState('10:00 AM');
 
   useEffect(() => {
     const privacyAccepted = localStorage.getItem('privacyAccepted_v1');
@@ -59,7 +58,6 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
 
   function handleChooseFulfillment(type) {
     setFulfillmentType(type);
-    setOrderType(type);
     setFulfillmentOpen(false);
     if (type === 'delivery') {
       setDeliveryModalOpen(true);
@@ -143,36 +141,32 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
     }));
   }, [state.items]);
 
-  // Order Type Selection Component
+  const cartTotal = getCartTotal();
+
   const OrderTypeSelection = () => (
     <div className="order-type-section">
       <div className="section-header">
         <h2>Order details</h2>
         <p>Select an order type</p>
       </div>
-      
       <div className="order-options">
         <div className="order-option">
           <h3>Vegetarian Slatter</h3>
           <p>11 products</p>
         </div>
-        
         <div className="order-option">
           <h3>Non-Veg Slatter</h3>
           <p>14 products</p>
         </div>
-        
         <div className="order-option">
           <h3>Soup</h3>
           <p>6 products</p>
         </div>
-        
         <div className="order-option">
           <h3>Main Vegetarian</h3>
           <p>28 products</p>
         </div>
       </div>
-      
       <div className="pickup-time">
         <h3>Pickup time</h3>
         <div className="time-option">
@@ -180,7 +174,6 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
           <span>{pickupTime}</span>
         </div>
       </div>
-      
       <div className="order-ready">
         <h3>ORDER READY FOR</h3>
         <div className="ready-time">
@@ -192,8 +185,7 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
   );
 
   return (
-    <div className="online-ordering-container">
-      {/* Backdrop for mobile drawer */}
+    <div>
       <div className="cart-backdrop hide-desktop" data-show={mobileCartOpen ? 'true' : 'false'} onClick={() => setMobileCartOpen(false)} />
 
       <CartSidebar
@@ -212,12 +204,11 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
       />
       <main className="content">
         <StoreHeader siteSlug={siteSlug} />
-        
-        {/* Order Type Selection Section */}
+
         <div className="card order-type-card">
           <OrderTypeSelection />
         </div>
-        
+
         <StoreBanner
           siteSlug={siteSlug}
           onCta={() => {
@@ -225,20 +216,14 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
             else if (state.fulfillmentType === 'delivery') setDeliveryModalOpen(true);
           }}
         />
-        
-        {selectedCategory && (
-          <>
-            <ShopTopBar vegFilter={vegFilter} onVegChange={setVegFilter} />
-            <section className="card" style={{ padding: 10, marginBottom: 10 }}>
-              <CategoryChips
-                categories={allCategories}
-                currentId={selectedCategory?._id}
-                onSelect={(c) => setSelectedCategory(c)}
-              />
-            </section>
-          </>
-        )}
-        
+        <ShopTopBar vegFilter={vegFilter} onVegChange={setVegFilter} />
+        <section className="card" style={{ padding: 10, marginBottom: 10 }}>
+          <CategoryChips
+            categories={allCategories}
+            currentId={selectedCategory?._id}
+            onSelect={(c) => setSelectedCategory(c)}
+          />
+        </section>
         <div className="hide-desktop" style={{ marginBottom: 10 }}>
           <button
             className="primary-btn"
@@ -256,16 +241,15 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
         {content}
       </main>
 
-      {/* Order Summary Bar (Fixed at bottom) */}
       <div className="order-summary-bar">
         <div className="order-empty">
           {state.items.length === 0 ? 'Your order is empty' : `${state.items.length} items in cart`}
         </div>
         <div className="order-total">
           <div className="total-label">Subtotal</div>
-          <div className="total-amount">${(state.total || 0).toFixed(2)}</div>
+          <div className="total-amount">${(cartTotal || 0).toFixed(2)}</div>
         </div>
-        <button 
+        <button
           className="confirm-btn"
           disabled={state.items.length === 0}
           onClick={() => {
@@ -281,7 +265,6 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
         </button>
       </div>
 
-      {/* Mobile FAB */}
       <button className="cart-fab hide-desktop" onClick={() => setMobileCartOpen(true)}>
         <span style={{ fontSize: 18 }}>🛒</span>
         <span style={{ fontWeight: 800 }}>Cart ({state.items.length})</span>
@@ -301,9 +284,7 @@ const Main = ({ siteSlug = 'default', initialCategoryId }) => {
       ) : null}
       <SpiceModal open={spiceOpen} spiceLevels={pendingProduct?.spiceLevels} onCancel={() => setSpiceOpen(false)} onConfirm={confirmSpice} />
       <ExtrasModal open={extrasOpen} groups={pendingProduct?.extraOptionGroups} onCancel={() => setExtrasOpen(false)} onConfirm={confirmExtras} />
-      {/* Toast for add-to-cart */}
       <AddToCartToast />
-      {/* Login moved to dedicated /login route */}
       <LoginModal open={loginOpen} onClose={() => setLoginOpen(false)} onSuccess={() => {
         setLoginOpen(false);
         if (!state.fulfillmentType) setFulfillmentOpen(true);
